packages/pytest-httpchain-templates/pyproject.toml
[project]
name = "pytest-httpchain-templates"
version = "0.1.0"
description = "Templating support for pytest-httpchain"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Alexander Eresov", email = "aeresov@gmail.com" }]
dependencies = [
    "pytest-httpchain-jsonref",
    "pytest-httpchain-models",
    "jsonref>=1.1.0",
    "pydantic>=2.11.0",
    "jmespath>=1.0.1",
    "requests>=2.32.4",
    "jsonschema>=4.0.0",
    "simpleeval>=1.0.3",
]

[build-system]
requires = ["uv_build>=0.7.21,<0.8.0"]
build-backend = "uv_build"
